+++ CHANGELOG.md
@@ -19,6 +19,11 @@
   - `vertical-embed`
     -  [#728](https://github.com/wix-incubator/rich-content/pull/728) Vertical Embed Plugin - alpha verison
 
+  - 'viewer'
+    - [#908](https://github.com/wix-incubator/rich-content/pull/908) Support Viewer predefined anchors
+
+  
+
 </details>
 <hr/>