+++ CHANGELOG.md
@@ -15,12 +15,15 @@
     Click to see more.
   </summary>
 
+### :rocket: New Feature
+- `fullscreen`
+  - [#845](https://github.com/wix-incubator/rich-content/pull/845) fullscreen image titles and design changes
 ### :bug: Bug Fix
 - `viewer`
   - [#929](https://github.com/wix-incubator/rich-content/pull/929) fix: empty lists viewer issues
 ### :house: Internal
 - `vertical-embed`
-  -  [#728](https://github.com/wix-incubator/rich-content/pull/728) Vertical Embed Plugin - alpha verison
+  - [#728](https://github.com/wix-incubator/rich-content/pull/728) Vertical Embed Plugin - alpha verison
 - 'viewer'
   - [#908](https://github.com/wix-incubator/rich-content/pull/908) Support Viewer predefined anchors